#437 - Move execution for IIFE to index.js
