+++ test/i18n.configureStaticCatalog.js
@@ -1,4 +1,4 @@
-var i18n = require('../i18n'),
+var i18n = require('../index'),
   should = require("should");
 
 describe('staticCatalog configuration', function() {
