+++ test/i18n.init.js
@@ -1,4 +1,4 @@
-var i18n = require('../i18n'),
+var i18n = require('../index'),
   should = require("should"),
   sinon = require("sinon"),
   fs = require('fs');
