+++ test/i18n.setup.js
@@ -1,4 +1,4 @@
-var i18n = require('../i18n'),
+var i18n = require('../index'),
     pkgVersion = require('../package.json').version,
     should = require("should");
