+++ test/i18n.verifyLocaleSelectionMethods.js
@@ -1,4 +1,4 @@
-var i18n = require('../i18n'),
+var i18n = require('../index'),
   should = require("should");
 
 describe('when configuring selected locale', function () {
